requirements.txt
defusedxml==0.4.1
Django==1.10.4
django-taggit==0.21.3
future==0.16.0
markdown2==2.3.2
oauthlib==2.0.1
Pillow==3.4.2
pycryptodomex==3.4.3
pyjwkest==1.3.2
PyJWT==1.4.2
python-social-auth==0.3.1
python3-openid==3.0.10
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
social-auth-core==0.1.0
